Comunicação com servidor funcionando: 1 - Verificando e-mail; 2 - Registrando máquina; 3 - Obtendo operações.

+++ MachineGuardianMonitorProtect/src/br/com/mrezzosoftware/machineguardianmonitorprotect/core/MyUser32.java
@@ -35,7 +35,7 @@ public interface MyUser32 extends Library {
     boolean TranslateMessage(WinUser.MSG msg);
     LRESULT DispatchMessage(WinUser.MSG msg);
     
-    boolean ExitWindowsEx(int flags, int reason);
+    boolean ExitWindowsEx(long flags, long reason);
     boolean LockWorkStation();
     
 }
